refactor(server): clean xmpp client boundaries (#883)
## Summary

Complete the staged XMPP client boundary cleanup in one PR.

## Stages completed

1. Split `waddle-xmpp-client-ffi` into focused boundary modules while
preserving the UniFFI public surface.
2. Added typed FFI/WASM send outcomes instead of collapsing expected
send failures into raw strings or rejected promise shapes.
3. Updated chat WASM send handling so retryable outcomes stay queued and
deterministic typed failures are removed from the retry queue.
4. Moved the large `waddle-xmpp-client` call test module out of
production code into `messaging/call/tests.rs`.
5. Added Apple UniFFI binding drift validation and normalized generated
Swift/header output.
6. Added Apple consumer CI that builds the XCFramework, macOS app, and
iOS app.
7. Aligned generated CI workflows with the branch's pinned cuenv
`0.51.0` release.

## XEP review

- Local `./xeps` is present and was reviewed before protocol-boundary
work.
- This PR does not add new XMPP wire semantics or custom namespaces.
- Existing XEP-owned behavior remains covered by the focused Rust/chat
tests touched here.

## Adversarial review

Review round findings addressed:

- FFI/Apple: drift-only validation did not prove Apple consumers; added
XCFramework plus macOS and iOS app build coverage.
- FFI/Apple: drift scripts did not use `--locked`; added locked cargo
invocations.
- FFI/Apple: generated bindings carried trailing whitespace; normalized
generation in both scripts.
- WASM/chat: typed non-`sent` outcomes were previously collapsed to null
sending ids; now they reject explicitly.
- WASM/chat: deterministic `invalid-recipient` / `invalid-options`
outcomes no longer remain in durable retry queues.
- CI: generated workflows installed cuenv `0.41.3` while the branch
required `0.51.0`; regenerated workflows now install `0.51.0`.

Second review note: the thread hit the local subagent limit before a
fresh post-fix subagent round could be spawned. I performed a local
adversarial pass over the fixes and am relying on CI/GitHub review as
the external post-push gate.

## Validation

- `cd server && cargo fmt --all -- --check`
- `cd server && cargo test -p waddle-xmpp-client-ffi`
- `cd server && cargo test -p waddle-xmpp-client-wasm`
- `cd server && cargo test -p waddle-xmpp-client` (rerun outside sandbox
for the local socket transport test)
- `cd server && cargo clippy -p waddle-xmpp-client -p
waddle-xmpp-client-ffi -p waddle-xmpp-client-wasm --all-targets -- -D
warnings`
- `bash server/scripts/check-xmpp-client-ffi-bindings.sh`
- `cd server && cuenv sync ci --check -A`
- `cd chat && bun test tests/client-send-readiness.test.ts
tests/offline-queue.test.ts`
- `cd chat && bun run lint`
- `cd chat && bun run build`
- `bash scripts/build-xcframework.sh --debug`
- `xcodebuild -project apps/apple/Waddle.xcodeproj -scheme Waddle-macOS
-destination 'platform=macOS' CODE_SIGNING_ALLOWED=NO build`
- `ruby -e 'require "yaml"; Dir[".github/workflows/*.yml"].each { |f|
YAML.load_file(f) }; puts "workflow yaml ok"'`
- `bash -n scripts/build-xcframework.sh
server/scripts/check-xmpp-client-ffi-bindings.sh`
- `git diff --check`

Local limitation:

- `xcodebuild -project apps/apple/Waddle.xcodeproj -scheme Waddle-iOS
-destination 'generic/platform=iOS Simulator' CODE_SIGNING_ALLOWED=NO
build` processed the iOS simulator XCFramework slice, then failed
locally because this machine's Xcode/iOS platform install is
inconsistent (`iOS 26.5 Platform Not Installed` / CoreSimulator version
mismatch). The CI Darwin runner is the authoritative iOS consumer gate.
