+++ .github/workflows/waddle-apple-pullrequest.yml
@@ -1,0 +1,34 @@
+name: waddle-apple-pullrequest
+
+on:
+  pull_request:
+    paths:
+      - .github/workflows/waddle-apple-pullrequest.yml
+      - apps/apple/**
+      - scripts/build-xcframework.sh
+      - server/Cargo.lock
+      - server/Cargo.toml
+      - server/crates/waddle-xmpp-client/**
+      - server/crates/waddle-xmpp-client-ffi/**
+      - server/rust-toolchain.toml
+  workflow_dispatch: {}
+
+concurrency:
+  group: ${{ github.workflow }}-${{ github.head_ref || github.ref }}
+  cancel-in-progress: true
+
+jobs:
+  apple-xmpp-client:
+    name: apple-xmpp-client
+    runs-on: namespace-profile-darwin-arm64
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+      - name: Install Rust targets
+        run: rustup target add aarch64-apple-darwin x86_64-apple-darwin aarch64-apple-ios aarch64-apple-ios-sim x86_64-apple-ios
+      - name: Build XMPP client XCFramework
+        run: bash scripts/build-xcframework.sh --debug
+      - name: Build macOS app
+        run: xcodebuild -project apps/apple/Waddle.xcodeproj -scheme Waddle-macOS -destination 'platform=macOS' CODE_SIGNING_ALLOWED=NO build
+      - name: Build iOS app
+        run: xcodebuild -project apps/apple/Waddle.xcodeproj -scheme Waddle-iOS -destination 'generic/platform=iOS Simulator' CODE_SIGNING_ALLOWED=NO build
