+++ ci/contributors/cuenv.cue
@@ -18,7 +18,7 @@ import "github.com/cuenv/cuenv/schema"
 			  aarch64|arm64) cuenv_asset="cuenv-linux-arm64" ;;
 			  *) echo "Unsupported Linux architecture: $arch" >&2; exit 1 ;;
 			esac
-			curl -fsSL -o /usr/local/bin/cuenv "https://github.com/cuenv/cuenv/releases/download/0.41.3/${cuenv_asset}" && chmod +x /usr/local/bin/cuenv && /usr/local/bin/cuenv sync -A
+			curl -fsSL -o /usr/local/bin/cuenv "https://github.com/cuenv/cuenv/releases/download/0.51.0/${cuenv_asset}" && chmod +x /usr/local/bin/cuenv && /usr/local/bin/cuenv sync -A
 			"""
 	}]
 }
