+++ server/wasm-pkg/waddle-xmpp-client-wasm/waddle_xmpp_client_wasm.js
@@ -1,7 +1,7 @@
 /* @ts-self-types="./waddle_xmpp_client_wasm.d.ts" */
-import wasmUrl from "./waddle_xmpp_client_wasm_bg.wasm?url&b=64eb083207c4";
-import * as bgModule from "./waddle_xmpp_client_wasm_bg.js?b=64eb083207c4";
-import { __wbg_set_wasm } from "./waddle_xmpp_client_wasm_bg.js?b=64eb083207c4";
+import wasmUrl from "./waddle_xmpp_client_wasm_bg.wasm?url&b=05694a19296d";
+import * as bgModule from "./waddle_xmpp_client_wasm_bg.js?b=05694a19296d";
+import { __wbg_set_wasm } from "./waddle_xmpp_client_wasm_bg.js?b=05694a19296d";
 
 let initPromise;
 
@@ -30,4 +30,4 @@ export default async function init() {
 // WaddleConfig, …) AND Rust free functions (xep0392_consistent_hue,
 // xep0392_consistent_color, …). A hand-curated list silently drops new
 // #[wasm_bindgen] free functions until somebody notices the chat crashing.
-export * from "./waddle_xmpp_client_wasm_bg.js?b=64eb083207c4";
+export * from "./waddle_xmpp_client_wasm_bg.js?b=05694a19296d";
